+++ packages/forest_admin_agent/lib/forest_admin_agent/routes/resources/native_query.rb
@@ -39,6 +39,7 @@ module ForestAdminAgent
           query.gsub!('?', args[:params][:record_id].to_s) if args[:params][:record_id]
           self.type = args[:params][:type]
           result = execute_query(
+            @datasource,
             query,
             args[:params][:connectionName],
             @permissions,