+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.12.9](https://github.com/ForestAdmin/agent-ruby/compare/v1.12.8...v1.12.9) (2025-10-27)
+
+
+### Bug Fixes
+
+* handle properly composite datasource ([#179](https://github.com/ForestAdmin/agent-ruby/issues/179)) ([b63f39a](https://github.com/ForestAdmin/agent-ruby/commit/b63f39ae38d7c088494b7abf9237d0597cce5be9))
+
+## [1.12.8](https://github.com/ForestAdmin/agent-ruby/compare/v1.12.7...v1.12.8) (2025-10-27)
+
+
+### Bug Fixes
+
+* **config:** add setting to disable route cache ([#189](https://github.com/ForestAdmin/agent-ruby/issues/189)) ([5acf768](https://github.com/ForestAdmin/agent-ruby/commit/5acf7686ca141c02facc2260181a45d13716e19c))
+
+## [1.12.7](https://github.com/ForestAdmin/agent-ruby/compare/v1.12.6...v1.12.7) (2025-10-27)
+
+
+### Bug Fixes
+
+* **filters:** operators now try to parse the value into the field's type before comparing the values ([#188](https://github.com/ForestAdmin/agent-ruby/issues/188)) ([b0dd9b9](https://github.com/ForestAdmin/agent-ruby/commit/b0dd9b9486932e5e49765c811c8011c6f0f606c8))
+
 ## [1.12.6](https://github.com/ForestAdmin/agent-ruby/compare/v1.12.5...v1.12.6) (2025-10-27)